Use unsigned long long for output formats

+++ fakes/include/codegen/FakeOutputFormat.h
@@ -32,10 +32,10 @@ namespace codegen
         public:
             FakeOutputFormat();
 
-            void write(unsigned char  data, Section section) override;
-            void write(unsigned short data, Section section) override;
-            void write(unsigned int   data, Section section) override;
-            void write(unsigned long  data, Section section) override;
+            void write(unsigned char      data, Section section) override;
+            void write(unsigned short     data, Section section) override;
+            void write(unsigned int       data, Section section) override;
+            void write(unsigned long long data, Section section) override;
 
             size_t getPosition(Section section) override;
             size_t getSectionStart(Section section) override;
